socket-http server v1 working

+++ ClientSocket.js
@@ -1,7 +1,8 @@
 var net = require('net');
-var str ="Hello, Im";
-
+var str ="quit";
 var client = new net.Socket();
+var cuenta = 0;
+
 client.connect(45000, '127.0.0.1', function() {
 
 	console.log('Connected');
@@ -10,11 +11,16 @@ client.connect(45000, '127.0.0.1', function() {
 
 client.on('data', function(data) {
 	console.log('Received: ' + data);
-client.write(str);
-
+	client.write(str);
+	if(cuenta>1)
 	client.destroy(); // kill client after server's response
+	cuenta++;
 });
 
 client.on('close', function() {
 	console.log('Connection closed');
 });
+
+
+
+
